requirements/base.txt
Django==1.8.5
django-debug-toolbar==1.3.2
Pillow==2.8.1
django-jfu==2.0.9
django-uuid-upload-path==1.0.0
django-resized==0.3.5
django-filter==0.11.0
django-vanilla-views==1.0.4
django-tables2==1.0.4
cookiecutter==1.2.1
django-compressor==2.0
numpy==1.10.4
scipy==0.17.0